Made MaxPlayers and ViewDistance default to -1 and ignore when not changed.
These parameters are type Integer and currently can't be left unset,
else the following error occurs on-launch:

Parameter 'ViewDistance' must be a number.
Parameter 'MaxPlayers' must be a number.

CloudFormation doesn't seem to have a simple solution for optional Integer
type parameters.  Setting to -1 and ignoring that seems reasonable.

Fixed #14.