Fixed possible loss of LaunchEvent processing from first ASG launch.
Prior to this change, the LaunchEvent, which captures the ASG instance
launch and processes a DNS change when configured, is depending on the
ASG itself, as it is referencing the instance in order to capture its
name.

This change reverses the dependency, by replacing the LaunchEvent's
reference from an instance to a string, which will match the ASG's
group name, and adds Metadata to the ASG, purely to create a depends-on
relationship.

Note that Metadata is used to create ASG depends-on (is created after)
the LaunchEvent, as it is not possible to define a conditional DependsOn
in CloudFormation, and LaunchEvent is not always defined if DNS is not
optionally configured.